Implemented ability to use any file extension. (Provided the expected extension contains only one dot).
Added a --require option to load in modules like 'coffee-script/register'. It has the alias -r and is treated as a comma separated list. These modules are loaded before other files to allow for non-javascript config files.

Also support a require setting in the config file for other modules that might be loaded globally. This is useful if the settings file is JSON.

Using test_settings.filter, it is now possible to filter by **/*.coffee etc to use a different file type. This is accomplished by setting filename_filter to default to **/*.js

Custom Commands, Assertions and Page Objects can be filtered by filename using the configuration setting custom_files_filter which defaults to *.js. Also, for the truly insane, each object can be filtered for individually using custom_commands_filter, custom_assertions_filter and page_objects_filter. The file extension is replaced by regex and only replaces beyond the last dot.

Updated filematcher to take an array of filters similar to the exclude behaviour.

Added coffee-script as a dev dependency.