Adds support for a global reporter function.
This provides a really simple way to hook in your own code or reporting
module. (see #54)

Adds support for a global test results reporter function defined in
the module configured in `globals_path` in `nightwatch.json`. A
function defined under the `reporter` key in the globals module is
called after all tests finish with the nightwatch internal results
object as an argument. It may be asynchronous, with a 'done' callback
as the reporter's second argument.

I recommend starting with a simple `console.log` (as seen in
examples/globalsModule.js) to view the test results data structure.