Fix silently swallowed exceptions
When something goes wrong and neither a test suite nor a finishCallback are given,
the processListener in run.js swallows the error. This happens when you are using
mocha as a test runner and have a syntax error or throw an exception outside of
the test suite (e.g. `require('./some-spec-helper')` while the file does not exist)