Priority given to configuration passed in as a string
To facilitate usage of nightwatch executions as part of server process that runs from js code, allow the user to pass in a stingified example of settings that would ordinarily be in nightwatch.json. This is highly dynamic and easier to debug since one can keep the logic that generates new configurations (say I want to change globals each time) in the node process and not the nightwatch process.