Enable passing arrow functions to `execute`
One of my colleagues noticed the interesting fact that they could pass
an arrow function with a block body and explicit return, but not an
arrow function with an expression body and an implicit return.

That is, this would work:
```
  protocol.execute(() => { return 'something'; });
```

While this would cause an exception:
```
  protocol.execute(() => 'something');
```

Given that these two functions are meant to be equivalent, this was a
very surprising result.

The underlying cause is that we serialize the function and concatenate
it with additional code that was expecting to be able to chain on to the
function. Without braces around the string representation, this
concatenation ends up trying to call `apply` on the result of the
expression in the function, instead of calling it on the function
itself.

We can resolve this by wrapping the serialized function in parenthesis,
disambiguating the call target for `apply`.

I haven't added an explicit test for this, because it would fail in the
Node 0.10 CI environment. :(