Add option for nightwatch.conf.js to be asynchronous
Added option of nightwatch config file exporting a function, which takes a single param - done. This allows for async settings files, with the settings file calling done() when finished (passing the settings). Adjusted sections of CLI runner to be more async and use callbacks to call functions to prevent nightwatch running ahead of the settings finishing.